refactor(core): split c/k/g spelling rule into syllable/spelling.rs
violates_ckg_spelling moves verbatim into the existing syllable/ dir,
bringing syllable.rs back under the 150-code-line budget (155 -> ~137).
No behavior change.
